Refactored DrawChart to simplify code assignment and modularity
Changed layout of code declarations so that the l_SVG is passed through the other functions that utilize the GitHub Chart and adjusted various lines for readability. Swapped table name link for button. I might make this a separate function in a utils.js file. Changed data types to better reflect scope of usage.